+++ packages/react-native/src/ThemeProvider/index.ts
@@ -1,0 +1,1 @@
+export { ThemeProvider } from './ThemeProvider';
