+++ packages/react-native/src/hooks/index.ts
@@ -2,4 +2,3 @@ export {
   DeviceOrientation,
   useDeviceOrientation,
 } from './useDeviceOrientation';
-export { useTheme } from './useTheme';
