+++ packages/react-native/src/theme/hooks/useTheme.ts
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
-import { ThemeContext } from '../ThemeProvider/ThemeContext';
-import { StrictTheme } from '../theme';
+import { ThemeContext } from '../../ThemeProvider/ThemeContext';
+import { StrictTheme } from '../../theme';
 
 export const useTheme = (): StrictTheme => {
   const context = React.useContext(ThemeContext);
